[FEAT] Split input fields from Component
